Use pom-scijava as parent
This makes the project independent on the rest of Fiji. Of course, the
groupId and the version tag need to be set now.

Also, the properties that were defined in pom-fiji need to be inserted,
too. This trick was performed by:

	for property in $(grep '\${' pom.xml |
		sed -e 's/.*\${//' -e 's/}.*//' |
		sort)
	do
		cat $FIJI/pom.xml |
		grep /$property
	done > pom.xml.properties
	sed -i.bak '/<properties>/rpom.xml.properties' pom.xml

The mpicbg.version had to be fixed up because there are still only release
versions reflecting the date when they were made, for lack of a statement
as of the intended next release version (which we later found out was to
be 0.6.0, which however was never committed, let alone tagged, released
nor deployed). See also

	https://github.com/axtimwalde/mpicbg/issues/6

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>