Add an exclusion to avoid Bio-Formats' testng dependency
Bio-Formats version 5.0.0 has a seriously broken pom.xml infrastructure,
by missing <scope>test</scope> entries (among a ton of other stuff).

This is the *bare* minimum fix to get Fiji compiling again, thanks to
TestNG's inability to interpret assumeNotNull() correctly (instead
assuming that it is the same as assertNotNull()).

A full work-around for Bio-Formats' borked POMs is in the making, thanks
to the tireless Curtis Rueden who cannot be thanked enough for all the
fixes he provides in world record time.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>