Update to use Bio-Formats 5.1 API
In particular, the sub-C API was removed, so we have to stop using it.
We could use the new Modulo[CZT] API instead, but it is more limited,
and it would be a substantial effort to adopt.

The number of use cases which:

A) Are still using ImgLib1; and
B) Use the ImgLib1 ImageOpener; and
C) Need the sub-C dimensions (e.g., SDT files)

is probably vanishingly small, so let's not worry about it till a
concrete need presents itself.